+++ example/lib/events/todo_added_event.dart
@@ -1,0 +1,7 @@
+import 'package:example/sdk/todo/proxies/todo.dart';
+import 'package:floater/floater.dart';
+
+class TodoAddedEvent extends Event {
+  final Todo todo;
+  TodoAddedEvent(this.todo);
+}