+++ example/lib/events/todo_updated_event.dart
@@ -1,0 +1,7 @@
+import 'package:example/sdk/todo/proxies/todo.dart';
+import 'package:floater/floater.dart';
+
+class TodoUpdatedEvent extends Event {
+  final Todo todo;
+  TodoUpdatedEvent(this.todo);
+}